Added Alarm and Toast-making Interface
Date picker now choses the time for an alarm and sends the data over
bluetooth in 24-hour format. Also, it is now possible to make toast from
any page in the app! In other news, the splash screen Somul logo looks a
little higher resolution and the light from the lamp is no longer cut
off on the side.